+++ deode/mars_utils.py
@@ -8,10 +8,36 @@ from pathlib import Path
 from subprocess import run
 from typing import Dict, List, Tuple
 
+from .config_parser import ParsedConfig
 from .datetime_utils import as_datetime
 from .geo_utils import Projection, Projstring
 from .logs import logger
 from .toolbox import Platform
+
+
+def mars_selection(selection: str, config: ParsedConfig) -> dict:
+    """Copy default settings if requested.
+
+    Args:
+        selection             (str): The selection to use.
+        config (deode.ParsedConfig): Configuration object
+
+    Returns:
+         mars                (dict): mars config section
+
+    """
+    mars = config[f"mars.{selection}"].dict()
+    if "expver" not in mars:
+        mars["expver"] = selection
+
+    # Copy default settings if requested
+    if "default" in mars:
+        default = config[f"mars.{mars['default']}"]
+        for k in default:
+            if k not in mars:
+                mars[k] = default[k]
+
+    return mars
 
 
 def write_retrieve_mars_req(req, name: str, method: str, omode: str = "w"):