+++ deode/tasks/boundaries.py
@@ -1,17 +1,16 @@
 """Boundary interpolation."""
 
-import ast
 import json
 import os
 from typing import Dict
 
+from deode.boundary_utils import Boundary
 from deode.config_parser import ConfigPaths
-from deode.datetime_utils import as_datetime, as_timedelta, cycle_offset
+from deode.datetime_utils import as_datetime
 from deode.logs import logger
 from deode.namelist import NamelistGenerator
 from deode.tasks.base import Task
 from deode.tasks.batch import BatchJob
-from deode.tasks.marsprep import Marsprep
 
 
 class InterpolateBoundaries(Task):
@@ -25,32 +24,14 @@ class InterpolateBoundaries(Task):
         """
         Task.__init__(self, config, __class__.__name__)
 
-        basetime = self.config["general.times.basetime"]
+        self.boundary = Boundary(config)
 
-        self.bd_index_time_dict = ast.literal_eval(
-            self.config.get("task.args.bd_index_time_dict")
-        )
-
-        bdmodel_map = {"lam": "e927", "ifs": "c903"}
-        self.bdmodel = self.config["boundaries.bdmodel"]
-        self.method = bdmodel_map[self.bdmodel]
-
-        if self.method == "e927":
-            bdcycle = as_timedelta(self.config["boundaries.lam.bdcycle"])
-            bdcycle_start = as_timedelta(self.config["boundaries.lam.bdcycle_start"])
-            self.namelists = {self.method: "fort.4"}
+        if self.boundary.method == "e927":
+            self.namelists = {self.boundary.method: "fort.4"}
             self.logs = ["fort.4", "NODE.001_01"]
             self.outfile = "PF@CNMEXP@000+0000"
 
-        elif self.method == "c903":
-            mars = Marsprep.mars_selection(
-                selection=self.platform.substitute(
-                    self.config[f"boundaries.{self.bdmodel}.selection"]
-                ),
-                config=self.config,
-            )
-            bdcycle = as_timedelta(mars["ifs_cycle_length"])
-            bdcycle_start = as_timedelta(mars["ifs_cycle_start"])
+        elif self.boundary.method == "c903":
             self.namelists = {
                 "c903_main": "fort.4",
                 "c903_domain": "namelist_c903_domain",
@@ -64,19 +45,14 @@ class InterpolateBoundaries(Task):
             + f"{self.config['file_templates.interpolated_boundaries.archive']}"
         )
 
-        # Boundary basetime
-        self.basetime = as_datetime(basetime)
-        bdshift = as_timedelta(config["boundaries.bdshift"])
-        self.bd_basetime = self.basetime - cycle_offset(
-            self.basetime, bdcycle, bdcycle_start=bdcycle_start, bdshift=bdshift
-        )
-
-        self.input_definition = f"{self.method}_input_definition"
+        self.input_definition = f"{self.boundary.method}_input_definition"
         self.nlgen = NamelistGenerator(self.config, "master")
-        self.master = self.get_binary("MASTERODB", task_name=self.method.upper())
-        self.min_index = min(self.bd_index_time_dict.keys())
-        self.max_index = max(self.bd_index_time_dict.keys())
-        self.name = f"{self.method}_{self.min_index}-{self.max_index}".upper()
+        self.master = self.get_binary("MASTERODB", task_name=self.boundary.method.upper())
+        self.name = (
+            f"{self.boundary.method}_"
+            f"{self.boundary.min_index}-"
+            f"{self.boundary.max_index}"
+        ).upper()
 
     def execute(self):
         """Run task.
@@ -97,7 +73,7 @@ class InterpolateBoundaries(Task):
             "host_model_files"
         )
 
-        if self.method == "c903":
+        if self.boundary.method == "c903":
             host_model_static_files: Dict[str, str | Dict[str, str]] = input_data.pop(
                 "host_model_static_files"
             )
@@ -109,16 +85,18 @@ class InterpolateBoundaries(Task):
                 # be able to reference files created by marsprep, which contains
                 # bdmember = 0 even for "deterministic" runs.
                 src_local = src
-                if not self.config.get(f"boundaries.{self.bdmodel}.bdmember"):
+                if not self.config.get(f"boundaries.{self.boundary.bdmodel}.bdmember"):
                     src_local = src_local.replace("@BDMEMBER@", "0")
                 self.fmanager.input(
                     f"{path_static}/{src_local}",
                     dst,
-                    basetime=self.bd_basetime,
-                    validtime=as_datetime(self.bd_index_time_dict[self.min_index]),
+                    basetime=self.boundary.bd_basetime,
+                    validtime=as_datetime(
+                        self.boundary.bd_index_time_dict[self.boundary.min_index]
+                    ),
                 )
         # Link the static data
-        self.fmanager.input_data_iterator(input_data, basetime=self.bd_basetime)
+        self.fmanager.input_data_iterator(input_data, basetime=self.boundary.bd_basetime)
 
         # Host model input files
         path = host_model_files["path"]
@@ -128,13 +106,13 @@ class InterpolateBoundaries(Task):
             # be able to reference files created by marsprep, which contains
             # bdmember = 0 even for "deterministic" runs.
             src_local = src
-            if not self.config.get(f"boundaries.{self.bdmodel}.bdmember"):
+            if not self.config.get(f"boundaries.{self.boundary.bdmodel}.bdmember"):
                 src_local = src_local.replace("@BDMEMBER@", "0")
-            for bd_index, bd_time in self.bd_index_time_dict.items():
+            for bd_index, bd_time in self.boundary.bd_index_time_dict.items():
                 self.fmanager.input(
                     f"{path}/{src_local}",
-                    dst.replace("@NNN@", f"{bd_index - self.min_index}"),
-                    basetime=self.bd_basetime,
+                    dst.replace("@NNN@", f"{bd_index - self.boundary.min_index}"),
+                    basetime=self.boundary.bd_basetime,
                     validtime=as_datetime(bd_time),
                 )
 
@@ -143,7 +121,7 @@ class InterpolateBoundaries(Task):
             if key == "c903_main":
                 self.nlgen.load(key)
                 self.nlgen.update(
-                    {"NAMCT0": {"NFRPOS": len(self.bd_index_time_dict)}},
+                    {"NAMCT0": {"NFRPOS": len(self.boundary.bd_index_time_dict)}},
                     "bd_dynamic_update",
                 )
                 nml = self.nlgen.assemble_namelist(key)
@@ -156,7 +134,7 @@ class InterpolateBoundaries(Task):
         batch.run(self.master)
 
         # Store result
-        for bd_index in self.bd_index_time_dict:
+        for bd_index in self.boundary.bd_index_time_dict:
             self.target = (
                 f"{self.platform.get_system_value('intp_bddir')}"
                 + "/"
@@ -164,7 +142,7 @@ class InterpolateBoundaries(Task):
             ).replace("@NNN@", f"{bd_index:03}")
 
             self.fmanager.output(
-                self.outfile.replace("@NNN@", f"{bd_index - self.min_index:04}"),
+                self.outfile.replace("@NNN@", f"{bd_index - self.boundary.min_index:04}"),
                 self.target,
             )
         self.archive_logs(self.logs)